Rename files, gem and test database to departure
