+++ lib/departure/errors.rb
