+++ lib/departure/logger_factory.rb
