+++ lib/departure/railtie.rb
@@ -1,10 +1,10 @@
-require 'percona_migrator'
+require 'departure'
 require 'lhm' # It's our own Lhm adapter, not the gem
 require 'rails'
 
 module PerconaMigrator
   class Railtie < Rails::Railtie
-    railtie_name :percona_migrator
+    railtie_name :departure
 
     # It drops all previous database connections and reconnects using this
     # PerconaAdapter. By doing this, all later ActiveRecord methods called in
@@ -13,13 +13,13 @@ module PerconaMigrator
     # It also patches ActiveRecord's #migrate method so that it patches LHM
     # first. This will make migrations written with LHM to go through the
     # regular Rails Migration DSL.
-    initializer 'percona_migrator.configure_rails_initialization' do
+    initializer 'departure.configure_rails_initialization' do
       ActiveSupport.on_load(:active_record) do
         PerconaMigrator.load
       end
     end
 
-    initializer 'percona_migrator.configure' do |app|
+    initializer 'departure.configure' do |app|
       PerconaMigrator.configure do |config|
         config.tmp_path = app.paths['tmp'].first
       end
