+++ spec/departure/cli_generator_spec.rb
