+++ spec/departure/null_logger_spec.rb
